feat: modify form feat: title paywall

+++ PetReminder/Repository/AnimalRepository.swift
@@ -60,9 +60,12 @@ final class AnimalRepository: AnimalRepositoryProtocol {
     func delete(_ animal: Animal) async throws {
         AnalitycsManager.shared.log(.deleteAnimal(AnimalAnalitycsEvent(animal: animal)))
         try await withThrowingTaskGroup(of: Void.self) { taskGroup in
-            taskGroup.addTask {
-                try await FirestoreService.deleteImage(animal.imagePath)
+            if !animal.imagePath.isEmpty {
+                taskGroup.addTask {
+                    try await FirestoreService.deleteImage(animal.imagePath)
+                }
             }
+            
             taskGroup.addTask {
                 try await FirestoreService.request(PetsEndpoints.deletePet(id: animal.id))
             }
